// src/types.ts
export interface FAQItem {
  id: string;
  question: string;
  answer: string;
}

export interface FleetCategory {
  id: string;
  name: string;
  tag: string;
  description: string;
  image: string;
}

export interface BizTab {
  id: string;
  title: string;
  subtitle: string;
  heading: string;
  description: string;
  image: string;
}

export interface AdvantageItem {
  id: string;
  title: string;
  description: string;
  image: string;
}

export interface FleetFacility {
  id: string;
  title: string;
  description: string;
  image: string;
}